PLATOPS-1326: Fix typo in package name
