+++ app/uk/gov/hmrc/servicedependencies/persistence/RepositoryLibraryDependenciesRepository.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.servicedependencies.presistence
+package uk.gov.hmrc.servicedependencies.persistence
 
 import com.google.inject.{Inject, Singleton}
 import org.joda.time.{DateTime, DateTimeZone}
