+++ app/uk/gov/hmrc/servicedependencies/service/DependencyDataUpdatingService.scala
@@ -24,9 +24,9 @@ import uk.gov.hmrc.lock.LockFormats.Lock
 import uk.gov.hmrc.servicedependencies.config.CuratedDependencyConfigProvider
 import uk.gov.hmrc.servicedependencies.controller.model.{Dependencies, Dependency}
 import uk.gov.hmrc.servicedependencies.model._
-import uk.gov.hmrc.servicedependencies.presistence._
+import uk.gov.hmrc.servicedependencies.persistence._
+import uk.gov.hmrc.servicedependencies.persistence._
 import uk.gov.hmrc.time.DateTimeUtils
-
 import scala.concurrent.Future
 
 @Singleton
