+++ test/uk/gov/hmrc/servicedependencies/persistence/LibraryVersionRepositorySpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.servicedependencies.presistence
+package uk.gov.hmrc.cataloguefrontend.events
 
 /*
  * Copyright 2017 HM Revenue & Customs
@@ -41,10 +41,11 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.modules.reactivemongo.ReactiveMongoComponent
 import uk.gov.hmrc.mongo.{MongoConnector, MongoSpecSupport}
 import uk.gov.hmrc.play.test.UnitSpec
-import uk.gov.hmrc.servicedependencies.model.{MongoSbtPluginVersion, Version}
+import uk.gov.hmrc.servicedependencies.model.{MongoLibraryVersion, Version}
+import uk.gov.hmrc.servicedependencies.persistence.LibraryVersionRepository
 import uk.gov.hmrc.time.DateTimeUtils
 
-class SbtPluginVersionRepositorySpec
+class LibraryVersionRepositorySpec
     extends UnitSpec
     with LoneElement
     with MongoSpecSupport
@@ -61,44 +62,44 @@ class SbtPluginVersionRepositorySpec
     override def mongoConnector = mockedMongoConnector
   }
 
-  val mongoSbtPluginVersions = new SbtPluginVersionRepository(reactiveMongoComponent)
+  val mongoLibraryVersions = new LibraryVersionRepository(reactiveMongoComponent)
 
   override def beforeEach() {
-    await(mongoSbtPluginVersions.drop)
+    await(mongoLibraryVersions.drop)
   }
 
   "update" should {
     "inserts correctly" in {
 
-      val sbtPluginVersion = MongoSbtPluginVersion("some-sbtPlugin", Some(Version(1, 0, 2)), DateTimeUtils.now)
-      await(mongoSbtPluginVersions.update(sbtPluginVersion))
+      val libraryVersion = MongoLibraryVersion("some-library", Some(Version(1, 0, 2)), DateTimeUtils.now)
+      await(mongoLibraryVersions.update(libraryVersion))
 
-      await(mongoSbtPluginVersions.getAllEntries) shouldBe Seq(sbtPluginVersion)
+      await(mongoLibraryVersions.getAllEntries) shouldBe Seq(libraryVersion)
     }
 
-    "updates correctly (based on sbtPlugin name)" in {
+    "updates correctly (based on library name)" in {
 
-      val sbtPluginVersion    = MongoSbtPluginVersion("some-sbtPlugin", Some(Version(1, 0, 2)), DateTimeUtils.now)
-      val newSbtPluginVersion = sbtPluginVersion.copy(version = Some(Version(1, 0, 5)))
-      await(mongoSbtPluginVersions.update(sbtPluginVersion))
+      val libraryVersion    = MongoLibraryVersion("some-library", Some(Version(1, 0, 2)), DateTimeUtils.now)
+      val newLibraryVersion = libraryVersion.copy(version = Some(Version(1, 0, 5)))
+      await(mongoLibraryVersions.update(libraryVersion))
 
-      await(mongoSbtPluginVersions.update(newSbtPluginVersion))
+      await(mongoLibraryVersions.update(newLibraryVersion))
 
-      await(mongoSbtPluginVersions.getAllEntries) shouldBe Seq(newSbtPluginVersion)
+      await(mongoLibraryVersions.getAllEntries) shouldBe Seq(newLibraryVersion)
     }
   }
 
   "clearAllDependencyEntries" should {
     "deletes everything" in {
 
-      val sbtPluginVersion = MongoSbtPluginVersion("some-sbtPlugin", Some(Version(1, 0, 2)), DateTimeUtils.now)
-      await(mongoSbtPluginVersions.update(sbtPluginVersion))
+      val libraryVersion = MongoLibraryVersion("some-library", Some(Version(1, 0, 2)), DateTimeUtils.now)
+      await(mongoLibraryVersions.update(libraryVersion))
 
-      await(mongoSbtPluginVersions.getAllEntries) should have size 1
+      await(mongoLibraryVersions.getAllEntries) should have size 1
 
-      await(mongoSbtPluginVersions.clearAllData)
+      await(mongoLibraryVersions.clearAllData)
 
-      await(mongoSbtPluginVersions.getAllEntries) shouldBe Nil
+      await(mongoLibraryVersions.getAllEntries) shouldBe Nil
     }
   }
 }
